+++ app/Werewolf/Slack/Slack.hs
@@ -26,8 +26,8 @@ import Network.HTTP.Types.Method
 
 import Werewolf.Slack.Options
 
-notify :: (MonadIO m, MonadReader Options m, MonadState Manager m) => String -> String -> m ()
-notify to message = do
+notify :: (MonadIO m, MonadReader Options m, MonadState Manager m) => Maybe String -> String -> m ()
+notify mTo message = do
     manager <- get
 
     initialRequest  <- asks optWebhookUrl >>= liftIO . parseUrl
@@ -41,6 +41,6 @@ notify to message = do
     where
         body    = RequestBodyLBS $ encode payload
         payload = object
-            [ "channel" .= to
+            [ "channel" .= mTo
             , "text" .= message
             ]